Add more configuration for MQTT
Add configuration details for MQTT, update tests as needed

Update makefile so testing is more verbose.

Add coverage option to makefile so that it's easy to view how the
testing is exercising the code.

Update program so that the formatting for the logger is specified
early so that it is consistent even if the config loading causes a
panic.